Fix: Compiler warning C4838: Conversion from 'uint32_t' to 'INT' requires a narrowing conversion.

+++ SDF/src/Graphics/Image.cpp
@@ -84,7 +84,7 @@ namespace sdf {
 
         hResult = decoderFrame->GetSize((UINT*)&width, (UINT*)&height);
 
-        WICRect rect = { 0, 0, width, height };
+        WICRect rect = { 0, 0, (INT)width, (INT)height };
 
         WICPixelFormatGUID pixelFormatGUID;
         hResult = decoderFrame->GetPixelFormat(&pixelFormatGUID);
